Updates to ultrasonic component and pic32 init files
Changed PBCLK freq divider to 1:2.
Updated ultrasonic module to calculate distance based on timer freq.

+++ Project_2_Ultrasonic.X/ping_ultrasonic.h
@@ -11,9 +11,9 @@
 void Ultrasonic_Init(void);
 void TriggerMeasurement(void);
 double GetDistance(void);
-uint32_t GetDistance_v2(void);
+double GetDistance_v2(void);
 double MeasureDistance(void);
-uint32_t MeasureDistance_v2(void);
+double MeasureDistance_v2(void);
 
 #endif	/* PING_ULTRASONIC_H */
